@@ -11,16 +11,16 @@ repository = "https://github.com/BurntSushi/ripgrep/tree/master/crates/searcher"
 readme = "README.md"
 keywords = ["regex", "grep", "egrep", "search", "pattern"]
 license = "Unlicense OR MIT"
-edition = "2018"
+edition = "2021"
 
 [dependencies]
-bstr = { version = "1.6.0", default-features = false, features = ["std"] }
-encoding_rs = "0.8.14"
-encoding_rs_io = "0.1.6"
+bstr = { version = "1.6.2", default-features = false, features = ["std"] }
+encoding_rs = "0.8.33"
+encoding_rs_io = "0.1.7"
 grep-matcher = { version = "0.1.6", path = "../matcher" }
-log = "0.4.5"
-memchr = "2.6.2"
-memmap = { package = "memmap2", version = "0.5.3" }
+log = "0.4.20"
+memchr = "2.6.3"
+memmap = { package = "memmap2", version = "0.8.0" }
 
 [dev-dependencies]
 grep-regex = { version = "0.1.11", path = "../regex" }
