@@ -13,24 +13,28 @@
 // with non-obvious failure modes. Alas, such things haven't been documented
 // well.
 
-use std::collections::HashMap;
-use std::ffi::{OsStr, OsString};
-use std::fs::{File, FileType};
-use std::io::{self, BufRead};
-use std::path::{Path, PathBuf};
-use std::sync::{Arc, RwLock};
+use std::{
+    collections::HashMap,
+    ffi::{OsStr, OsString},
+    fs::{File, FileType},
+    io::{self, BufRead},
+    path::{Path, PathBuf},
+    sync::{Arc, RwLock},
+};
 
-use crate::gitignore::{self, Gitignore, GitignoreBuilder};
-use crate::overrides::{self, Override};
-use crate::pathutil::{is_hidden, strip_prefix};
-use crate::types::{self, Types};
-use crate::walk::DirEntry;
-use crate::{Error, Match, PartialErrorBuilder};
+use crate::{
+    gitignore::{self, Gitignore, GitignoreBuilder},
+    overrides::{self, Override},
+    pathutil::{is_hidden, strip_prefix},
+    types::{self, Types},
+    walk::DirEntry,
+    {Error, Match, PartialErrorBuilder},
+};
 
 /// IgnoreMatch represents information about where a match came from when using
 /// the `Ignore` matcher.
 #[derive(Clone, Debug)]
-pub struct IgnoreMatch<'a>(IgnoreMatchInner<'a>);
+pub(crate) struct IgnoreMatch<'a>(IgnoreMatchInner<'a>);
 
 /// IgnoreMatchInner describes precisely where the match information came from.
 /// This is private to allow expansion to more matchers in the future.
@@ -85,7 +89,7 @@ struct IgnoreOptions {
 
 /// Ignore is a matcher useful for recursively walking one or more directories.
 #[derive(Clone, Debug)]
-pub struct Ignore(Arc<IgnoreInner>);
+pub(crate) struct Ignore(Arc<IgnoreInner>);
 
 #[derive(Clone, Debug)]
 struct IgnoreInner {
@@ -134,22 +138,22 @@ struct IgnoreInner {
 
 impl Ignore {
     /// Return the directory path of this matcher.
-    pub fn path(&self) -> &Path {
+    pub(crate) fn path(&self) -> &Path {
         &self.0.dir
     }
 
     /// Return true if this matcher has no parent.
-    pub fn is_root(&self) -> bool {
+    pub(crate) fn is_root(&self) -> bool {
         self.0.parent.is_none()
     }
 
     /// Returns true if this matcher was added via the `add_parents` method.
-    pub fn is_absolute_parent(&self) -> bool {
+    pub(crate) fn is_absolute_parent(&self) -> bool {
         self.0.is_absolute_parent
     }
 
     /// Return this matcher's parent, if one exists.
-    pub fn parent(&self) -> Option<Ignore> {
+    pub(crate) fn parent(&self) -> Option<Ignore> {
         self.0.parent.clone()
     }
 
@@ -157,7 +161,7 @@ impl Ignore {
     ///
     /// Note that this can only be called on an `Ignore` matcher with no
     /// parents (i.e., `is_root` returns `true`). This will panic otherwise.
-    pub fn add_parents<P: AsRef<Path>>(
+    pub(crate) fn add_parents<P: AsRef<Path>>(
         &self,
         path: P,
     ) -> (Ignore, Option<Error>) {
@@ -222,7 +226,7 @@ impl Ignore {
     /// returned if it exists.
     ///
     /// Note that all I/O errors are completely ignored.
-    pub fn add_child<P: AsRef<Path>>(
+    pub(crate) fn add_child<P: AsRef<Path>>(
         &self,
         dir: P,
     ) -> (Ignore, Option<Error>) {
@@ -335,7 +339,7 @@ impl Ignore {
     }
 
     /// Like `matched`, but works with a directory entry instead.
-    pub fn matched_dir_entry<'a>(
+    pub(crate) fn matched_dir_entry<'a>(
         &'a self,
         dent: &DirEntry,
     ) -> Match<IgnoreMatch<'a>> {
@@ -520,7 +524,7 @@ impl Ignore {
     }
 
     /// Returns an iterator over parent ignore matchers, including this one.
-    pub fn parents(&self) -> Parents<'_> {
+    pub(crate) fn parents(&self) -> Parents<'_> {
         Parents(Some(self))
     }
 
@@ -534,7 +538,7 @@ impl Ignore {
 /// An iterator over all parents of an ignore matcher, including itself.
 ///
 /// The lifetime `'a` refers to the lifetime of the initial `Ignore` matcher.
-pub struct Parents<'a>(Option<&'a Ignore>);
+pub(crate) struct Parents<'a>(Option<&'a Ignore>);
 
 impl<'a> Iterator for Parents<'a> {
     type Item = &'a Ignore;
@@ -552,7 +556,7 @@ impl<'a> Iterator for Parents<'a> {
 
 /// A builder for creating an Ignore matcher.
 #[derive(Clone, Debug)]
-pub struct IgnoreBuilder {
+pub(crate) struct IgnoreBuilder {
     /// The root directory path for this ignore matcher.
     dir: PathBuf,
     /// An override matcher (default is empty).
@@ -572,7 +576,7 @@ impl IgnoreBuilder {
     ///
     /// All relative file paths are resolved with respect to the current
     /// working directory.
-    pub fn new() -> IgnoreBuilder {
+    pub(crate) fn new() -> IgnoreBuilder {
         IgnoreBuilder {
             dir: Path::new("").to_path_buf(),
             overrides: Arc::new(Override::empty()),
@@ -596,7 +600,7 @@ impl IgnoreBuilder {
     ///
     /// The matcher returned won't match anything until ignore rules from
     /// directories are added to it.
-    pub fn build(&self) -> Ignore {
+    pub(crate) fn build(&self) -> Ignore {
         let git_global_matcher = if !self.opts.git_global {
             Gitignore::empty()
         } else {
@@ -638,7 +642,10 @@ impl IgnoreBuilder {
     /// By default, no override matcher is used.
     ///
     /// This overrides any previous setting.
-    pub fn overrides(&mut self, overrides: Override) -> &mut IgnoreBuilder {
+    pub(crate) fn overrides(
+        &mut self,
+        overrides: Override,
+    ) -> &mut IgnoreBuilder {
         self.overrides = Arc::new(overrides);
         self
     }
@@ -648,13 +655,13 @@ impl IgnoreBuilder {
     /// By default, no file type matcher is used.
     ///
     /// This overrides any previous setting.
-    pub fn types(&mut self, types: Types) -> &mut IgnoreBuilder {
+    pub(crate) fn types(&mut self, types: Types) -> &mut IgnoreBuilder {
         self.types = Arc::new(types);
         self
     }
 
     /// Adds a new global ignore matcher from the ignore file path given.
-    pub fn add_ignore(&mut self, ig: Gitignore) -> &mut IgnoreBuilder {
+    pub(crate) fn add_ignore(&mut self, ig: Gitignore) -> &mut IgnoreBuilder {
         self.explicit_ignores.push(ig);
         self
     }
@@ -665,7 +672,7 @@ impl IgnoreBuilder {
     ///
     /// When specifying multiple names, earlier names have lower precedence than
     /// later names.
-    pub fn add_custom_ignore_filename<S: AsRef<OsStr>>(
+    pub(crate) fn add_custom_ignore_filename<S: AsRef<OsStr>>(
         &mut self,
         file_name: S,
     ) -> &mut IgnoreBuilder {
@@ -676,7 +683,7 @@ impl IgnoreBuilder {
     /// Enables ignoring hidden files.
     ///
     /// This is enabled by default.
-    pub fn hidden(&mut self, yes: bool) -> &mut IgnoreBuilder {
+    pub(crate) fn hidden(&mut self, yes: bool) -> &mut IgnoreBuilder {
         self.opts.hidden = yes;
         self
     }
@@ -687,7 +694,7 @@ impl IgnoreBuilder {
     /// supported by search tools such as ripgrep and The Silver Searcher.
     ///
     /// This is enabled by default.
-    pub fn ignore(&mut self, yes: bool) -> &mut IgnoreBuilder {
+    pub(crate) fn ignore(&mut self, yes: bool) -> &mut IgnoreBuilder {
         self.opts.ignore = yes;
         self
     }
@@ -698,7 +705,7 @@ impl IgnoreBuilder {
     /// file path given are respected. Otherwise, they are ignored.
     ///
     /// This is enabled by default.
-    pub fn parents(&mut self, yes: bool) -> &mut IgnoreBuilder {
+    pub(crate) fn parents(&mut self, yes: bool) -> &mut IgnoreBuilder {
         self.opts.parents = yes;
         self
     }
@@ -711,7 +718,7 @@ impl IgnoreBuilder {
     /// This overwrites any previous global gitignore setting.
     ///
     /// This is enabled by default.
-    pub fn git_global(&mut self, yes: bool) -> &mut IgnoreBuilder {
+    pub(crate) fn git_global(&mut self, yes: bool) -> &mut IgnoreBuilder {
         self.opts.git_global = yes;
         self
     }
@@ -722,7 +729,7 @@ impl IgnoreBuilder {
     /// man page.
     ///
     /// This is enabled by default.
-    pub fn git_ignore(&mut self, yes: bool) -> &mut IgnoreBuilder {
+    pub(crate) fn git_ignore(&mut self, yes: bool) -> &mut IgnoreBuilder {
         self.opts.git_ignore = yes;
         self
     }
@@ -733,7 +740,7 @@ impl IgnoreBuilder {
     /// `gitignore` man page.
     ///
     /// This is enabled by default.
-    pub fn git_exclude(&mut self, yes: bool) -> &mut IgnoreBuilder {
+    pub(crate) fn git_exclude(&mut self, yes: bool) -> &mut IgnoreBuilder {
         self.opts.git_exclude = yes;
         self
     }
@@ -743,7 +750,7 @@ impl IgnoreBuilder {
     ///
     /// When disabled, git-related ignore rules are applied even when searching
     /// outside a git repository.
-    pub fn require_git(&mut self, yes: bool) -> &mut IgnoreBuilder {
+    pub(crate) fn require_git(&mut self, yes: bool) -> &mut IgnoreBuilder {
         self.opts.require_git = yes;
         self
     }
@@ -751,7 +758,7 @@ impl IgnoreBuilder {
     /// Process ignore files case insensitively
     ///
     /// This is disabled by default.
-    pub fn ignore_case_insensitive(
+    pub(crate) fn ignore_case_insensitive(
         &mut self,
         yes: bool,
     ) -> &mut IgnoreBuilder {
@@ -768,7 +775,7 @@ impl IgnoreBuilder {
 /// precedence than later names).
 ///
 /// I/O errors are ignored.
-pub fn create_gitignore<T: AsRef<OsStr>>(
+pub(crate) fn create_gitignore<T: AsRef<OsStr>>(
     dir: &Path,
     dir_for_ignorefile: &Path,
     names: &[T],
@@ -861,22 +868,19 @@ fn resolve_git_commondir(
 
 #[cfg(test)]
 mod tests {
-    use std::fs::{self, File};
-    use std::io::Write;
-    use std::path::Path;
+    use std::{io::Write, path::Path};
 
-    use crate::dir::IgnoreBuilder;
-    use crate::gitignore::Gitignore;
-    use crate::tests::TempDir;
-    use crate::Error;
+    use crate::{
+        dir::IgnoreBuilder, gitignore::Gitignore, tests::TempDir, Error,
+    };
 
     fn wfile<P: AsRef<Path>>(path: P, contents: &str) {
-        let mut file = File::create(path).unwrap();
+        let mut file = std::fs::File::create(path).unwrap();
         file.write_all(contents.as_bytes()).unwrap();
     }
 
     fn mkdirp<P: AsRef<Path>>(path: P) {
-        fs::create_dir_all(path).unwrap();
+        std::fs::create_dir_all(path).unwrap();
     }
 
     fn partial(err: Error) -> Vec<Error> {
@@ -1193,7 +1197,7 @@ mod tests {
         assert!(ignore.matched("ignore_me", false).is_ignore());
 
         // missing commondir file
-        assert!(fs::remove_file(commondir_path()).is_ok());
+        assert!(std::fs::remove_file(commondir_path()).is_ok());
         let (_, err) = ib.add_child(td.path().join("linked-worktree"));
         // We squash the error in this case, because it occurs in repositories
         // that are not linked worktrees but have submodules.
