@@ -5,9 +5,12 @@ An implementation of `grep-matcher`'s `Matcher` trait for
 
 #![deny(missing_docs)]
 
-pub use crate::error::{Error, ErrorKind};
-pub use crate::matcher::{RegexCaptures, RegexMatcher, RegexMatcherBuilder};
 pub use pcre2::{is_jit_available, version};
+
+pub use crate::{
+    error::{Error, ErrorKind},
+    matcher::{RegexCaptures, RegexMatcher, RegexMatcherBuilder},
+};
 
 mod error;
 mod matcher;
