@@ -1,5 +1,4 @@
-use std::fs::File;
-use std::path::Path;
+use std::{fs::File, path::Path};
 
 use memmap::Mmap;
 
